set maximum line length to 99
PEP 8 has the following to say about the subject of line length:

"Some teams strongly prefer a longer line length. For code maintained
exclusively or primarily by a team that can reach agreement on this
issue, it is okay to increase the nominal line length from 80 to 100
characters (effectively increasing the maximum length to 99 characters),
provided that comments and docstrings are still wrapped at 72
characters."

As discussed in #13, we've settled on 99 as a compromise between
sticking to PEP8 and making use of modern screen widths.

Update the editorconfig and add a flake8 configuration file to the root
of the repo.