tox: customise output locations from environment variables
Currently we run tox as a special case in the developer docker-compose
with a writable volume. This is sub-optimal as it can leave root-owned
files scattered around. As a first step towards fixing this, allow the
three main outputs to be configured by environment variables. These are:

1. The tox work dir itself. Overridden with the TOXINI_WORK_DIR
   environment variable.
2. The build artefact dir. Overridden with the TOXINI_ARTEFACT_DIR
   environment variable.
3. The coverage data file. Overridden with the COVERAGE_FILE environment
   variable. (The coverage tool itself knows about this environment
   variable.)

Tidy up the tox.ini to avoid the needless "setup.py sdist install" step.
We do not use setup.py when deploying so we need not use it in tests.