docker-compose: fix bind volume mount bug
In other docker-compose projects we have encountered a similar error to
this error which I have just encountered:

> ERROR: Named volume "{u'read_only': True, u'source': '.', u'type':
> 'volume', u'target': '/usr/src/app'}" is used in service "devserver"
> but no declaration was found in the volumes section.

We have found that explicitly noting that those volumes are *bind*
volumes fixes this problem.

This commit applies the same fix to our docker-compose.yml file which
was (possibly) one of the earliest written and does not have this fix.