return 403 responses from unauthorised embed requests
Previously we returned a 200 response if a request was made to embed a
video without the appropriate ACLs and rendered this response by the
non-matchingly-named template "401.html".

In fact, neither 200 nor 401 responses are appropriate. The former
indicates a success (which it isn't) and a 401 request MUST include a
WWW-Authenticate header[1] (which it doesn't).

The correct response code is 403 Forbidden which indicates that the
server understood the response, could determine which user (if any) was
logged in and decided that the user cannot have the resource.

[1] http://tools.ietf.org/html/rfc7235#section-4.1