refactor ACL support to avoid double-fetch of video metadata
Previously, the get_acl() function would re-fetch video metadata which
we must have previously fetched to find the video in the first place.
This introduced the additional latency of a network call to the view.

Refactor the ACL support to be implemented by a dedicated Video object
which is a sub-class of the dictionary object. The Video object now has
from_media_id() and from_key() class methods which retrieve the
appropriate resource as before but, since we keep the full video
resource, we can now generate the ACLs without further network access.