Split docker-compose into separate files
In code-review it was noted that it would be nice to add the production
Docker image to docker-compose.yml. This would have made the
docker-compose.yml file extremely overloaded.

Instead, take the opportunity to re-factor the docker-compose.yml file
into separate files and use docker-compose's ability to use multiple
files[1] to separate the configuration into base configuration,
development configuration and production configuration.

Rather than relying on a tired developer to remember to include both
files, the docker-compose files can be used via a new ./compose.sh
script. E.g.:

  $ ./compose.sh development up -d  # launch devserver in background
  $ ./compose.sh tox run --rm tox   # run test suite
  $ ./compose.sh production up -d   # launch production in background

A docker-compose configuration for production has been included.
Importantly, it doesn't do the mount the local directory inside the
container trick we do for development so 1) you know that you are
actually testing the contents of the container and 2) you have to
remember to run ./compose.sh production build.

Update the documentation to match the new compose files and way of
running them.

Finally, the Travis configuration has been updated to run the test suite
within the container and to provide a basic check that the production
container launches and the health-check endpoint comes up.

The convention for Google-y projects is for the healt-check endpoint to
live under /healthz so we take the opportunity to move /status to
/healthz at the same time.

[1] https://docs.docker.com/compose/extends/