connect the frontend to the backend
This commit attempts to solve a somewhat gnarly problem: the frontend is
written with React and wants to have its own "development" server with
hot-reload and the backend is written in Django which also wants its own
"development" server. In production, do we base the image on the python
3.6 image or the node one? Etc, etc.

The solution in this commit is to try and have our cake and eat it: we
would like http://localhost:8000/ to be the Django development server
but have it serve the frontend. http://localhost:3000/ is the
live-reloading frontend served by create-react-app but it needs to be
able to use the Django API.

The easiest requirement to satisfy is that the frontend at
http://localhost:3000/ works. We simply add a new service to the
docker-compose configuration which runs "npm start" and exposes port
3000 to the host. We make use of create-react-app's built in proxying
support to proxy requests to the API through to the Django development
app.

Of course in production, we want the frontend to be served by the Django
app and we'd like to be able to test this in development. To allow
Django to serve the app, we add some whitenoise configuration which
arranges to serve files from a frontend build directory (configured via
a setting) from /. If a request to /foo can be satisfied by a file from
this directory, it is so otherwise the Django app takes a turn to
satisfy the request.

We add a new docker-compose service which makes use of the watch script
we added to the frontend to watch for changes and build a new bundle of
HTML/JavaScript in a docker volume. This same volume is mounted in the
Django app container and is the directory whitenoise looks at.

Finally, we come to production. The whitenoise configuration above is
intended for use in production as well. We arrange for the frontend
production build files to be in a directory in the production container
and tell whitenoise to serve files from it.

The subtlety is in how those files get there.

We want the production image to remain small and so installing the
entirety of node, npm and the (frankly, stupidly large) node_modules
directory seems troublesome since we'll never actually use them when
running the container. Instead we make use of a useful but little-known
feature in Docker called a multi-stage build[1]. With this feature a
Dockerfile can actually use multiple stages and only keep the last
stage. We use this feature to compile the frontend JavaScript inside a
node container and then copy the compiled JavaScript *out* of that
container and *in* to the actual production container. This has the
strong advantage that all of the node-related stuff then gets thrown
away after the image is built.

[1] https://docs.docker.com/develop/develop-images/multistage-build/