frontend: implement basic media and collectionsearch support
This commit is somewhat large in that it provides an initial
implementation of the frontend.

A lot of the content of this commit is in the form of components to
implement the UI. Documentation for these can be read via styleguidist
which also allows for interactive experimentation.

The main "weight" of the UI implementation is in the IndexPage component
which fetches the latest media items, connects the search form to the
search results request and lays out the basic page structure.

The main "active" code is in api.ts. This module implements an API
client and is documented and checked via TypeScript. We're not really
using TypeScript in anger here but it has the nice side-effect of
allowing us to document the API resource types in code rather than
relying on comments. Once the data escapes the TypeScript world, we lose
all of the advantages, of course.

The last non-trivial bit of code is the SearchResultsProvider component.
This component makes use of the new React context API to decouple the
fetching of search results from rendering them. (The ProfileProvider
does the same for the user profile.)

In the IndexPage, a SearchResultsProvider is created and given the
current search query via a query prop. Other components may "subscribe"
to the search results by means of the withSearchResults higher-order
component which will map the current query, results and an isLoading
flag into the wrapped component.

This is very much a "first pass" UI and there will no doubt be some
issues which need addressing as we move first.

Older browser support is provided by a) polyfill.io and b) adding the
browser-update.org banner to chide/warn users of ancient browsers that
they are probably in for a bad time.

It has been tested in IE11 but no earlier browser.