hydra: create OAuth2 clients at docker-compuse up time
Previously we required that developers remember to run the
create-client.sh script manually when deploying for the first time. This
seems a little silly when we could just call it for them. Do so by
creating a new docker-compose service which runs a tweaked
client-create.sh script within a new hydra container.

The client creation script will delete any existing clients so it is
safe to run it multiple times.

Add a new lookupproxy client to the client creation script to allow use
of the lookupproxy Swagger UI we serve.