work-around older docker-compose on CircleCI
The default docker-compose installed on the CircleCI machines does not
support version 3.4 of the docker-compose format. Unfortunately this
means that we cannot use extension-fields to template the frontend
services and must copy-paste them. Leave the template configuration in
there as a comment as an extended "TODO" to re-instate when CircleCI
updates.