move to using secrets.env in development
Experience elsewhere with the PI Dashboard has shown it is useful to
have docker-compose refuse to start if the appropriate secrets are not
in place.

Update the docker-compose development and production configurations to
load JWPlatform secrets from a "secrets.env" file rather than
teleporting them from the environment when starting the services.

Update documentation and README to note the new mechanism of running.