smsjwplatform: add CachedResource model
Add a table to cache JWPlatform resources. We take the simple approach:
each resource is recorded in a JSONB column with the resource type and
key recorded along with it. The key is used as the primary key of the
database.

Add a custom object manager with a helper method which takes an iterable
of key, resource dict pairs and updates the cache for a given resource
type. Under the covers this uses some SQL magic to minimise the number
of round-trips to the database. Essentially, we make the observation
that the database is best placed to know if it already has a copy of the
resource or not so we use the appropriate INSERT ... ON CONFLICT runes
to let PostgreSQL sort out itself what to do.

Overall, the resource cache update method manages to do a cache
population and expiry without incurring a round-trip delay for each
INSERT. This is all hidden inside the implementation, of course, so
callers need only know *what* it does, not the dirty arcane magical way
in which it does it.

Should Django develop upsert support (as appears to be in progress),
then we can replace the implementation as and when it appears.

Using something like django-postgres-extra would also solve our problem
but that's quite invasive. It involves using a different database
engine, for example, which might make things difficult for us re
deployment. We *could* stop the engine being configurable in the Docker
image but that's probably beyond the scope of this story and would
require further discussion.

[1] http://django-postgres-extra.readthedocs.io/manager/#conflict-handling