IndexPage: adjust spacing
Adjust the body spacing in the index page to make sections provide their
own vertical spacing so that the body area itself need no hard-coded
padding at the top. This is to allow banner content at the top of the
body which needs to sit hard up against the top edge.