add redirect view to legacy SMS support for handling media pages
Add an endpoint under /legacy for handling individual media pages. It
maps the SMS media id to a JWPlayer key and redirects to the new media
page. If the media item could not be found or if the ACL disallows
viewing, the user is redirected back to the original SMS application to
let it deal with the problem.

Closes #82