speed up first run of tox in container
By default, tox will create isolated virtualenvs which ensure that your
requirements file specifies all of the packages you need because
packages you've installed outside of the virtualenvs are not used.

This is great when running tox locally on your own machine but when
running tox in a container it is counter-productive since the container
*already* isolates you from your local environment.

Provide an escape hatch for indicating that you want to use the
sitepackages to tox by means of an environment variable and use that
variable in to tox docker-compose configuration. This means that the
initial creation of virtualenvs can re-use the existing packages
installed into the environment.