port model changes informed by trying to use the models
Actually trying to use the models has resulted in some lessons in how
things should be laid out. Particularly, it turns out to be the wrong
answer to have the {view,edit}_permission fields on the MediaItem model.
If this is done, there's no way of deleting the dependent permission
automatically when the MediaItem is deleted and the admin becomes
unusable when it tries to populate a dropdown with every permission in
the database.