mediaplatform.models: re-work ACL support
Re-work the permission check code to be agnostic as to whether it is
referring to view_permission or edit_permission. Add another helper
which will annotate a queryset with a boolean value indicating if the
specified user can view the video.

This is in preparation for adding edit permission checks which will use
this same plumbing.