api: add MediaItemPermission
Add a DRF permission which controls access to MediaItem objects. The
basic idea is that a view permission is sufficient to perform any "safe"
HTTP methods but the edit permission is required *in addition to the
view permission* to perform any "unsafe" methods.

Add an additional MediaItemEditPermission which requires that the
current user has the edit permission in all cases.