api: re-work media item views
Add in support for creating ann updating media items to the Media views.
Since this functionality should not be exposed to the ui view, re-work
the views to be based on a mixin class which provides all of the
queryset/permissions/etc specialisations shared between the API and UI
views and create some concrete views using the mixin based on DRF's
generic views.

This has the nice side-effect that our hack-ish override of the get()
method for the UI view is no longer required.