mediaplatform_jwp: bugfix - synchronisation was not respecting updated
The code for updating JWP videos was confusing the updated_at timestamp
for the MediaItem with the updated timestamp for the JWP video.
Unfortunately, the slack built in to the updater was masking this in the
test suite since all videos were being updated, irrespective of their
updated timestamp.

Fix this by explicitly keeping track of which JWP videos we modify and
using that list to decide which media items need synchronising. Not only
does this make the sync logic simpler, it also makes it correct.

Fixing the sync logic exposed some unit tests which make changes but do
not modify the updated timestamp and so, unsurprisingly, they were
failing. Fix those unit tests.