frontend: new component: ItemMetadataForm
Add a new component which edits a Media Item resource. It is a
controlled component so the component which contains it is responsible
for listening to the onChange event and updating the item
appropriately.

This is a minimal form which will definitely need extending in future.
For the moment, it edits the title and description.

The example shows a minimal "wrapping" component which maintains a draft
item as some state.