split api and JSON-LD rendering
This commit splits the serialiser used for the API from the serialiser
used for JSON-LD. The idea behind this is that JSON-LD is not rich
enough for our API (e.g. to express the legacy SMS link). Converting our
models to JSON-LD is also a little against the grain of the DRF.

So, instead we use the standard DRF ModelSerialisers for the API which
reduces the amount of custom code by quite some amount. We lightly
customise them so that the API looks "JavaScript-y" with camel case.

For JSON-LD, we move the serialiser into the ui application and
"generisise" it a bit so that it can be used for other resources in the
fullness of time.

We embed the JSON-LD into the page as before (but this time via a Django
filter which is probably the way we should've done it to begin with).

To keep the desired behaviour that we do not need an extra round trip to
the API, we also embed the API resource into the page and teach
the frontend how to find embedded API response. We do this in a fairly
generic way, however, so we can make use of it going forward.