CR: MediaPage: change styling of download and analytics link
To finish off #119, freshen the look of the links at the botom of the
MediaPage and tweak their spacing/sizing to be more firendly to mobile
devices.

In keeping with the latest Material Design guidelines, these buttons,
not being in a card but also not representing primary *actions* (as
opposed to navigation links), are rendered with the "outlined" style.