port to Django 2.1 and Python 2.7
Update the portions of our development and production packaging which
specify Django 2.0 and Python 3.6 to talk about Django 2.1 and Python
3.7. Bump the minimum required version of django-ucamwebauth to be
Django 2.1 compatible.

Make tox neutral on the precise version of Python it uses (beyond
Python3) since we always run tox inside a container and it should just
use whatever PYthon 3 is given to it.