api: add channels
Expose the Channel object in the API. This makes use of the generic
permission and list mixin base classes to provided access with very
little code.

The channel detail view itself includes a URL which points to the list
of media items for that channel. This is generated by forming the
absolute URL for the media list with the channel filter parameters
applied.