CR: jwpfetch: always synchronise channels
It was determined in testing that, annoyingly, JWP channel resources
have no updated timestamp so we cannot use them to filter the number of
updates to only those needing it. Fortunately, the number of channels is
small w.r.t. the number of videos so we can survive, for the moment,
updating all the resources.

If this becomes an issue, we can move to using the sms_last_updated_at
timestamp.