api: add channel as a required parameter to media item resource
Add a channel to the media item resource. When returned from the media
list, the channel is represented via a channel resource. When creating
or updating a media item, the channel is set via a channelId parameter.

With this change, a user *must* specify a valid channel for the media
item when creating it. The media item channel also cannot be changed
once set. We may want to allow users to move media items between
channels in future but for the moment disallow it since it would
complicate things.