frontend: add side drawer
Add a side navigation drawer to the application. The draw is response;
on small screens it is hidden behind a toggle menu button and on large
screens it is always visible. This is consistent with the recommendation
for applications in the Material Design Guidelines.

Again, following the Design Guidelines, replace the profile button with
a nice avatar + display name style panel.