fix legacy redirect
The redirect from /legacy/media/{media_id} to /media/{pk} was broken
because it assumed we were still using JWP keys as the primary keys for
videos. Fix the redirect and update test suite. The basic redirect test
now actually tries to fetch the redirect to determine if the redirect
location succeeds.

Closes #176