add Django robots.txt and sitemap framework
Add django-robots and django.contrib.sitemaps applications and provide
some configuration for the sitemaps.

The sitemap is configured at the moment to include the 200 most recently
updated media items and the index page. As we add channel views, etc, we
should add them to the sitemap to help Search Engines out.

The actual configuration of the robots.txt is via the Djano sites
framework and is performed in the admin so, once deployed, there is a
chore to go and configure all the sites.

The sites framework records the domain name of the site in the database.
The default post_migration hook creates an "example.com" entry which
also needs modifying in the admin on all the deployed sites.

Closes #173