api: add dedicated endpoint for poster image
Taking a leaf out of the sources endpoint book, add a dedicated API view
for the poster image. This allows us to hide JWP URLs so that the
backend can change but most users won't notice.

This also allows us to customise the width of the poster images in a
more documented way via a width query to the poster image endpoint. JWP
only supports a small set of poster images, however, to update the
swagger definition and endpoint implementation to know about this.