api: add embed view
Add an embed view for media items which redirect to the appropriate JWP
embed view. This should really be abstracted into the specific JWP
backend but that's a job for when we tidy up the backend/middleend
split.

Adding a dedicated embed view means that the embed URL returned for a
media item is now part of the API and is long-lived rather than being
the short-lives JWP embed URLs.

The embed view is special in that we need to prompt the user to log in
if the media item is not found since it may not be found only because
the item is protected. Because of this, we add a custom 404 handler for
embed view. The embed 404 view is based on the classic SMS one but
updated to use flexbox for positioning and to be black on white text
with Roboto. The sign in link is shown only if the user was not already
signed in.