api: add media source view
Add a specific view which will redirect to a media source and update the
list of sources returned by the API to point to them. The sources are
retrieved from the MediaItem via get_sources.

The source view itself matches the source based on mime type and (if
appropriate) width and height. The first matching source is always
returned.