CR: update palette
The secondary colour inhereted from the IAR was red and, in the words
of our UX expert, "only ever use red for severe or bad actions". Use the
material UI palette design tool along with the "blessed" University
colours to try and find a palette which is both based on the official
palette and also satisfies the Material accessibility requirements.