frontend: Page: remove sign in link from app bad
The sign in link is removed and the padding of the main content element
is modified so that an (optional) gutter can be automatically placed at
the top of the page. Pages which may not want this gutter include the
MediaItemPage which wants to have the video hard-up against the top of
the page.