frontend: fix material UI warnings about unknown classes
In StaticTextPage, we were attempting to pass the classes prop down to
the wrapped Paper element but Maerial-UI complained about unknown
styles. Fix this by having a "componentProps" prop on BodySection which
allows passing props directly to the wrapped component.

Update the appearance of the static text page on small and large
screens. On the former, remove the shadow which was overlapping the
alpha bar and on the latter add a margin to the bottom f the page so the
paper does not end exactly at the end of the page.