Move all jwplatform api calls under the jwplatformapi directory
We have two files for jwplatform apis, one for the delivery API and another one for the management api. Now they are both located in the jwplatformapi app which makes sense as the app dedicated to manage jwp. Move them to be in a folder toguether so it's easy to reference them and know where they are.