enable versioning for media items and channels
Use the django-reversion package to enable versioning for the media item
and channel objects. This gets logging and delete protection "for free".
After this has bedded down for a while, we can probably remove the
"deleted_at" fields from MediaItem and Channel.

This commit uses the django-reversion middleware to wrap *every* non
GET/HEAD/OPTIONS request. This is probably a little heavy handed but
makes sure we catch all of the ways we could mutate an object via the
UI/API.

It's not currently clear how the django-reversion module can be used
with any non Django session authentication used by DRF. Until we add
such authentication, this is a non-issue but it should be fairly easy to
create a version of the view decorator
(https://django-reversion.readthedocs.io/en/stable/views.html) which
works with DRF as a mixin class much like reversion.views.RevisionMixin.
Unlike the RevisionMixin class, the DRF one would wrap the dispatch()
method of a DRF view and wrap non GET/HEAD/OPTIONS requests with a
revision.

This has the advantage that we can be a bit more "surgical" in which
views we revision.