api: add mediaCount field to channel detail response
Add a count of media to the detail response for a channel. We place this
in the detail response because it involves running an aggregation over
all items in the channel which may be expensive to do when listing the
channels.

The media count respects the user making the request and returns only
the count of media items viewable by that user.

Required to implement the summaries in #249.