frontend: update design for NavigationPanel
Add some design elements from #249 to the navidation drawer. The "Latest
Media" list item is replaced with a "Home" entry, a feedback link is
added along with a list of the user's channels.

The "Sign in" link has been moved to the last group of items for
consistency with the "Sign out" link but is at the top of the group.

The channels list does not include the "Create channel" item since we do
not yet have a view for this.

The About, Contact and Developers links have been added to the bottom of
the navigation panel.

This provides the bulk of the non-trivial functionality in #249 but does
not fully implement the design, as noted above, since some features are
not yet present in the application.