add celery task runner to application
Add the celery task runner to the application and add configuration to
docker-compose so that a task runner is brought up in development.

We use the django-celery-beat application to allow scheduling of
periodic tasks via database models rather than configuration. The idea
behind this is that we have flexibility in whether we configure periodic
tasks via settings, explicitly via code in post-migration hooks or via
per-deployment database fixtures.

We add a simple "debug" task (called mediawebapp.debug) which simply
logs the request and its arguments.