CR: fix regression where new media items are not synced
In the rare case of a SMS derived media item not existing but the
mediaplatform_jwp.Video item existing *and* the CachedResource item
existing, the MediaItem object is created but its content is not
synchronised.

Fix the logic used to determine which media items need to be
resynchronised to include any items whose JWP-side metadata has changed
OR any media items which were created.

Add a test which covers this case.