mediaplatform: respect the mediaplatform.view_mediaitem permission
Update the viewable condition for a media item to include all users who
have the mediaplatform.view_mediaitem permission in the set of those who
can view a media item.

Make sure to actually call the _editable_condition() method within
_viewable_condition() so that a related change there (perhaps to do with
mediaplatform.change_mediaitem) will be reflected.

Add a test to check that a new user with no permissions gains access to
a media item when the permission is granted.