mediaplatform_jwp: associate {Video,Channel} with CachedResource
When reviewing #302, it was noted that it would be useful to be able to
directly access the resource associated with a particular JWP video or
channel rather than having to explicitly issue a separate query.

This commit adds that plumbing. We add a new non-NULL field to the
Channel and Video models which points to the associated resource and fix
up all the places where we create the models.

We use a custom migration to perform the following steps for videos and
channels:

1. Add the resource column but initially allow it to be NULL.
2. Fix up the resource links by querying the related object from the db.
3. Assert that no values are left as NULL.
4. Alter the resource column to disallow NULL.

The vast majority of this commit is fixing up the tests to understand
that JWP videos now need a CachedResource.