frontend: re-jig page layout
Re-jig the page layout so that the content of the page is not confined
to an overflow element within a flex box but is actually the full page.
This works around a nasty interaction between white-space:
nowrap/text-overflow: ellipsis and flexbox which means that the media
items were sometimes overflowing to the right of the page.

It also has the advantage that the Firefox screenshow "save full page"
option now correctly renders the full height of the page.

Closes #309