mediaplatform: anonymous users can never edit media items
As noted in #336, the anonymous media list page is far slower than it
really should be. That issue contains some digging on the matter but the
executive summary is that querying for related SMS objects in the edit
condition is very slow.

The anonymous user should never have an edit permission on anything so,
as suggested in #336, a quick win is to shortcut the edit condition
check for anonymous users.

This does not completely fix the problem in that non-anonymous views are
still slower than they should be.