api: fix incorrect merge
Commits 5a55d421 and aedbd5a5 attempted to generalise the way that
channel objects are annotated for ChannelDetailSerializer.
Unfortunately, they generalised it in different ways which meant that
the generalisation was only applied to the channels rendered as part of
the user profile.

Move the ProfileView over to inheriting from ViewMixinBase and teach the
latter how to query the current user's profile.

This new method is used both by ProfileView and all of the views in the
UI since the profile is rendered as part of each resource page. While we
are refactoring, remove the erroneous non-trivial work which is done in
the ResourcePageSerializer and place it in the view where it belongs.

The entire UI view/API view interaction is becoming increasingly
complex and brittle. We should re-visit the entire way the UI is
rendered since we don't *need* to render a corresponding resource or the
user profile into the page; we can simply use the API. The only
non-trivial view should be the media view since that needs to render
structured JSON for SEO purposes.

Closes #339