frontend: Snackbar: show snackbar messages immediately
Previously we would set a value in browser localStorage to be displayed
with the page is next loaded. This worked when we would always refresh
the entire page on each action but now we need to show the snackbar
message immediately.

Keep the basic interface the same: a function exported from the Snackbar
module. The name has been changed to reflect that the message is shown
immediately. We take the opportunity to base the Snackbar component on
the example from the Material UI docs.

Rather than selecting and implementing a flux framework for managing
this tiny bit of global state, we do it ourselves in the Snackbar
module. This is a bit HACK-y and should we move to a more formal
solution it should be refactored.

Fix up the various uses of snackbar messages away from redirecting via
window.location to show the snackbar to using the <Redirect> component.