api: add custom paginator for counting resources
Add a custom DRF paginator which is based on CursorPaginator byt can
optionally add a count of the number of results. This is disabled by
default because doing a full count can be inefficient and isn't always
needed.

Add appropriate magic to make sure that the Swagger API documentation
reflects the new change.