api: add billing account
Add billing account to the API. For the moment this stops channel
creation because there are precisely no billing accounts which a user
has channel create permission on because this permission has not been
created. This commit anticipates the coming of this change by adding the
plumbing necessary to protect channel billingAccountId from being
changed to a value corresponding to a billing account the user has no
rights on.

Add tests for the new api functionality.