oaipmh: MatterhornRecord: add concept of Series
Add a Series object to the database which records new series when they
are harvested. Each Series may have an associated playlist which is
intended to be where new videos for that series are added. The adding of
videos to playlists is not added by this commit.

When a MatterhornRecord is updated, the associated Series object is
ensured.

This commit will require a full cleanup to be applied to ensure that
Series objects are created for existing records.