oaipmh: create MediaItems for Tracks which need them
Add a new module, oaipmh.tracks, with a function,
ensure_track_media_item, which will create a media item for a Track if
necessary.

"Necessary" here means 1) it has a series, 2) that series has a playlist
and 3) the track does not already have a media item.

This makes use of the new "fetch_url" field for media items.

Ordinarily this function is called when a track updates via a post_save
hook. To catch when that fails (e.g. because the JWP managemenr API is
broken), we also extend the cleanup task so that the function is called
for tasks which need it.