mediaplatform_jwp: support a "fetch URL" for creating media items
The JWP API does not allow one to specify a download URL once the media
item has been created. This is annoying in the extreme.

For the moment, work around this by creating a new "fetch_url" field on
MediaItem. If this is non-blank, it is used to populate the
"download_url" parameter when creating the resource via the JWP API.