api: add test for media detail returned by PLaylist
The Playlist fixture did not actually have any real media items in it so
the list of media being returned from the API was always empty. Fix this
and add a test that, if all media in a playlist is public, all the media
items which exist are returned in the detail view.