api: filter returned media items by visibility
The existing playlist API would return details on *all* media in the
playlist even if the current user could not view it. The playlist object
always returns all media ids irrespective of visibility because a naked
media id is fairly innocuous. A full media item, however, includes
sensitive information such as title, description and poster image.

Filter the returned media and add a test which ensures that invisible
media does not appear.

Closes #398