oaipmh: allow per-series configuration of track types
The OAIPMH_TRACK_TYPES setting now controls the default track types for
new Series objects. The series object is updated to store a list of
track types for that Series which should be converted to media items.

The logic for track creation now uses the track type in the series to
determine which tracks to create.

Take the opportunity to change the default track type to
composite/delivery.