mediaplatform_jwp: tag and add custom props to created/updated resources
In order to be able to track on JWP which site a video is available on
and under which media id, we add some tags and custom props.

1. We add a tag of the form "sitedomain:some.domain.invalid" taken from
   the default Django Site whenever we update a JWP video.

2. We add two custom props prefixed by "site.{safe_domain}.". These
   props are "name" (the friendly name from the Django site
   configuration) and "media_id" (the media id of the item on the site.)

Here "safe_domain" is the site's domain name with "."s replaced by "_"s.
This is due to JWP interpreting "." specially within custom props. By
prefixing the custom prop with the site domain it is possible for the
same JWP video to appear on multiple sites without them fighting for
control.