frontend: new provider: GenericJWPConfigurationProvider
Add a new provider which is a generic implementation of the logic
required to turn the information from the UI-specific JWPlayer
configuration endpoints into a set of options which can be passed to
jwplayer(...).setup().

The data munging required is a little bit subtle so the provider is
heavily documented to describe exactly what it does.