frontend: new component: ResponsiveJWPlayerPlayer
Add a new component which provides styling around the base
JWPlayerProvider.Player component. This styling is intended to ensure
that the player stays at a 16:9 ratio and is centred within the page.

The CSS in this is deep and mysterious. Sorry. I've attempted to
document the basic principles behind it but I'd be lying if I claimed
there wasn't a fair bit of Googling crystalised into this code.