mediaplatform_jwp: proxy player library
As noted in #407, we were signing player.js URLs incorrectly. Implement the
legacy signing method for player URLs as documented at [1]. Looking at
the signing method, there are a number of weaknesses:

1. MD5 is a broken hash.

2. The secret is appended to the value to be hashed meaning that the
   hash state preceding the secret is known. HMAC, for example, guards
   against this explicitly.

To avoid exposing the signatures to all and sundry, we proxy the player
library so that the signature only goes between our server and JWP. We
cache the resulting response to avoid making a round-trip delay to JWP
be the common case and instead to make the common case that the player
JavaScript is served from the cache.

[1] https://developer.jwplayer.com/jw-platform/docs/developer-guide/delivery-api/legacy-url-token-signing/

Closes #407