frontend: PlaylistOverlay: fix indexing bug
As noted in #408, the logic used to highlight the current media item and
to select a new one in the embedded playlist view was faulty when the
playlist was being searched.

The root cause was the fact that the index of the item being rendered in
the playlist overlay must match the index of the item in the JWP
playlist itself. Filtering the list of items beforehand by the search
breaks this invariant. We fix this by imply rendering a null element for
playlist items which should not be rendered.

Closes #408